chore[apsystems]: remove old timekeeping code
Due too unknown reasons, the cronjob timekeeping was not working
correclty. Therefore too many request were send to the device which
might crash it, forcing a powerreset.

To prevent this, code was added to ensure only 1 request is send every
minute. Cronjob timekeeping seemed to work again, so old can be removed.